Move token select from placement modal to SDL builder (#63)
